Begin adapt Terms for batch whitelist :warning: Not tested. Commit to be able to switch branch
