+++ app/helpers/analyzer.py
@@ -1,7 +1,5 @@
 import abc
-from collections import defaultdict
 from configparser import NoOptionError
-from typing import DefaultDict, Optional, Dict
 
 import dateutil
 import copy
@@ -219,22 +217,3 @@ class Analyzer(abc.ABC):
     @abc.abstractmethod
     def evaluate_model(self):
         raise NotImplementedError()
-
-    @staticmethod
-    def add_term_to_batch(eval_terms_array: DefaultDict, aggregator_value: Optional[str], target_value: Optional[str],
-                          observations: Dict, doc: Dict) -> DefaultDict:
-        if aggregator_value not in eval_terms_array.keys():
-            eval_terms_array[aggregator_value] = defaultdict(list)
-
-        eval_terms_array[aggregator_value]["targets"].append(target_value)
-        eval_terms_array[aggregator_value]["observations"].append(observations)
-        eval_terms_array[aggregator_value]["raw_docs"].append(doc)
-
-        return eval_terms_array
-
-    @staticmethod
-    def remove_term_to_batch(eval_terms_array, aggregator_value, term_counter):
-        eval_terms_array[aggregator_value]["targets"].pop(term_counter)
-        eval_terms_array[aggregator_value]["observations"].pop(term_counter)
-        eval_terms_array[aggregator_value]["raw_docs"].pop(term_counter)
-        return eval_terms_array
